requirements.txt
# Core dependencies
cryptography>=41.0.0
PyJWT>=2.8.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
requests>=2.31.0
urllib3>=2.0.7

# Database
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.9
mysql-connector-python>=8.1.0

# Security
bcrypt>=4.0.1
python-jose>=3.3.0
passlib>=1.7.4
argon2-cffi>=23.1.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.1

# Development
black>=23.7.0
flake8>=6.1.0
mypy>=1.5.1
isort>=5.12.0

# Documentation
sphinx>=7.1.2
sphinx-rtd-theme>=1.3.0

# Monitoring
prometheus-client>=0.17.1
statsd>=4.0.1

# Logging
structlog>=23.1.0
python-json-logger>=2.0.7

# API
fastapi>=0.103.0
uvicorn>=0.23.2
pydantic>=2.3.0

# CLI
click>=8.1.7
rich>=13.5.2

# Utilities
python-dateutil>=2.8.2
pytz>=2023.3
tzlocal>=5.0.1

# Security scanning
tree-sitter>=0.20.0
# Temporarily removed syft and prowler as they have complex dependencies

# Cloud providers
boto3>=1.26.0
azure-mgmt-security>=3.0.0
google-cloud-securitycenter>=1.20.0

# Container security
docker>=6.0.0
kubernetes>=26.1.0

# AI/ML
torch>=2.0.0
transformers>=4.30.0
scikit-learn>=1.2.0
pandas>=2.0.0
numpy>=1.24.0

# Monitoring and notifications
watchdog>=3.0.0
slack-sdk>=3.20.0
schedule>=1.2.0
psutil>=5.9.0

# Development
pre-commit>=3.3.0

# Documentation
mkdocs>=1.4.0
mkdocs-material>=9.1.0

# API and Web
jinja2>=3.1.0
weasyprint>=59.0 